refactor(evals): move baseline harness into tests
